Seperated files into global variables, main, etc.
Added in couple of new functions that lets user set robot location or direction
