+++ LineTracking/global_vars.c
@@ -1,0 +1,10 @@
+/*
+ * global_vars.c
+ * Implementation of "global vars.h"
+ *
+ * Defines variables needed in various functions
+ */ 
+#include "global_vars.h"
+
+const char *NESW[] = {"North", "East", "South", "West"};
+//					  0        1       2        3
