+++ LineTracking/prompt_user.h
@@ -1,0 +1,16 @@
+/*
+ * prompt_user.h
+ * Declares functions that prompt the user to set robot values.
+ */ 
+#include "global_vars.h"
+
+#ifndef PROMPT_USER_H_
+#define PROMPT_USER_H_
+
+// Prompts user to set robot direction
+void prompt_set_direction();
+
+// Prompts user to set robot location
+void prompt_set_location();
+
+#endif /* PROMPT_USER_H_ */
